feat(test-runner): improve browser close logic

+++ packages/test-runner-browserstack/src/browserstackManager.ts
@@ -13,32 +13,46 @@ let connection: browserstack.Local | undefined = undefined;
 
 export const localId = `web-test-runner-${uuid()}`;
 
+async function setupLocalConnection(password: string, options: Partial<browserstack.Options> = {}) {
+  process.on('SIGINT', closeLocalConnection);
+  process.on('SIGTERM', closeLocalConnection);
+  process.on('beforeExit', closeLocalConnection);
+  process.on('exit', closeLocalConnection);
+
+  connection = new browserstack.Local();
+
+  console.log('[Browserstack] Setting up Browserstack Local proxy...');
+  await promisify(connection.start).bind(connection)({
+    key: password,
+    force: true,
+    localIdentifier: localId,
+    ...options,
+  });
+}
+
+function closeLocalConnection() {
+  if (connection && (connection as any).pid != null) {
+    process.kill((connection as any).pid);
+    connection = undefined;
+  }
+}
+
 export async function registerBrowserstackLocal(
   launcher: BrowserLauncher,
   password: string,
   options: Partial<browserstack.Options> = {},
 ) {
-  if (!connection) {
-    connection = new browserstack.Local();
-
-    console.log('[Browserstack] Setting up Browserstack Local proxy...\n');
-    await promisify(connection.start).bind(connection)({
-      key: password,
-      force: true,
-      localIdentifier: localId,
-      ...options,
-    });
-  }
-
   launchers.add(launcher);
+
+  if (!connection) {
+    await setupLocalConnection(password, options);
+  }
 }
 
 export function unregisterBrowserstackLocal(launcher: BrowserLauncher) {
   launchers.delete(launcher);
 
   if (connection && launchers.size === 0) {
-    if ((connection as any).pid != null) {
-      process.kill((connection as any).pid);
-    }
+    closeLocalConnection();
   }
 }
